[Refactor] 加入資金設定資料，完善提款頁面、輸入判斷

+++ app/src/main/java/org/cxct/sportlottery/network/money/MoneyRechCfg.kt
@@ -62,7 +62,7 @@ class MoneyRechCfg {
 
     class WithdrawCfg {
         var wdRate: Double? = null//用户提现费率
-        var withDrawBalanceLimit: Int? = null//每笔最小提现金额
+        var withDrawBalanceLimit: Long? = null//每笔最小提现金额
         var maxWithdrawMoney: Long? = null//每笔最高提现金额
     }
 }
